# Conflicts:
#	api/models/GameCourse/Module/Module.php
#	api/modules/Badges/Badges.php
#	api/modules/Leaderboard/Leaderboard.php
#	frontend/src/app/_services/api/api-http.service.ts
#	frontend/src/app/_views/restricted/courses/course/settings/roles/roles.component.html
#	frontend/src/app/_views/restricted/courses/course/settings/views/views/views.component.scss
#	frontend/src/app/shared.module.ts

+++ api/models/GameCourse/Course/Course.php
@@ -235,7 +235,7 @@ class Course
         // Validate data
         if (key_exists("name", $fieldValues)) {
             $newName = $fieldValues["name"];
-            self::validateName($newName);
+            self::validateName($newName, key_exists("year", $fieldValues) ? $fieldValues["year"] : $this->getYear(), $this->id);
             $oldName = $this->getName();
         }
         if (key_exists("short", $fieldValues)) self::validateShort($fieldValues["short"]);
@@ -539,23 +539,53 @@ class Course
      */
     public static function deleteCourse(int $courseId)
     {
-        // Remove data folder
-        Course::removeDataFolder($courseId);
+        $course = Course::getCourseById($courseId);
+        if ($course) {
+            // Disable modules
+            // NOTE: module dependants are disabled before the module
+            $modulesEnabled = $course->getModules(true, true);
+            $modulesDisabled = [];
+            while (count($modulesDisabled) != count($modulesEnabled)) {
+                foreach ($modulesEnabled as $moduleId) {
+                    if (in_array($moduleId, $modulesDisabled)) // already disabled
+                        continue;
 
-        // Disable autogame & remove info
-        AutoGame::setAutoGame($courseId, false);
-        AutoGame::deleteAutoGameInfo($courseId);
+                    // Check if dependants have already been disabled
+                    $allDependantsDisabled = true;
+                    foreach (($course->getModuleById($moduleId))->getDependants(DependencyMode::HARD, true) as $dependantId) {
+                        $dependant = $course->getModuleById($dependantId);
+                        if ($dependant->isEnabled()) {
+                            $allDependantsDisabled = false;
+                            break;
+                        }
+                    }
 
-        // Remove automations
-        $course = new Course($courseId);
-        $course->setAutomation("AutoEnabling", null);
-        $course->setAutomation("AutoDisabling", null);
+                    // Disable module
+                    if ($allDependantsDisabled) {
+                        $module = $course->getModuleById($moduleId);
+                        $module->setEnabled(false);
+                        $modulesDisabled[] = $moduleId;
+                    }
+                }
+            }
 
-        // Delete cache
-        Cache::clean($courseId);
+            // Remove data folder
+            Course::removeDataFolder($courseId);
 
-        // Delete from database
-        Core::database()->delete(self::TABLE_COURSE, ["id" => $courseId]);
+            // Disable autogame & remove info
+            AutoGame::setAutoGame($courseId, false);
+            AutoGame::deleteAutoGameInfo($courseId);
+
+            // Remove automations
+            $course->setAutomation("AutoEnabling", null);
+            $course->setAutomation("AutoDisabling", null);
+
+            // Delete cache
+            Cache::clean($courseId);
+
+            // Delete from database
+            Core::database()->delete(self::TABLE_COURSE, ["id" => $courseId]);
+        }
     }
 
     /**
@@ -664,7 +694,11 @@ class Course
             "u.*, a.username, a.auth_service, a.lastLogin, cu.lastActivity, cu.isActive as isActiveInCourse",
             "u.id"
         );
-        foreach ($courseUsers as &$courseUser) { $courseUser = CourseUser::parse($courseUser); }
+        foreach ($courseUsers as &$courseUser) {
+            $cu = CourseUser::getCourseUserById($courseUser["id"], $this);
+            $courseUser["image"] = $cu->getImage();
+            $courseUser = CourseUser::parse($courseUser);
+        }
         return $courseUsers;
     }
 
@@ -696,7 +730,11 @@ class Course
             "u.*, a.username, a.auth_service, a.lastLogin, cu.lastActivity, cu.isActive as isActiveInCourse",
             "u.id"
         );
-        foreach ($courseUsers as &$courseUser) { $courseUser = CourseUser::parse($courseUser); }
+        foreach ($courseUsers as &$courseUser) {
+            $cu = CourseUser::getCourseUserById($courseUser["id"], $this);
+            $courseUser["image"] = $cu->getImage();
+            $courseUser = CourseUser::parse($courseUser);
+        }
         return $courseUsers;
     }
 
@@ -743,7 +781,11 @@ class Course
             "u.*, a.username, a.auth_service, a.lastLogin",
             "u.id"
         );
-        foreach ($users as &$user) { $user = User::parse($user); }
+        foreach ($users as &$user) {
+            $u = User::getUserById($user["id"]);
+            $user["image"] = $u->getImage();
+            $user = User::parse($user);
+        }
         return $users;
     }
 
@@ -998,7 +1040,7 @@ class Course
      */
     public function getPages(?bool $visible = null): array
     {
-        return Page::getPagesOfCourse($this->id, $visible);
+        return Page::getPages($this->id, $visible);
     }
 
     /**
@@ -1228,7 +1270,7 @@ class Course
      */
     public function setAutoEnabling(?string $startDate)
     {
-        $script = ROOT_PATH . "models/GameCourse/Course/AutoEnablingScript.php";
+        $script = ROOT_PATH . "models/GameCourse/Course/scripts/AutoEnablingScript.php";
         if ($startDate) {
             // Check if date is in the future
             if (date("Y-m-d H:i:s", strtotime($startDate)) > date("Y-m-d H:i:s", time()))
@@ -1246,7 +1288,7 @@ class Course
      */
     public function setAutoDisabling(?string $endDate)
     {
-        $script = ROOT_PATH . "models/GameCourse/Course/AutoDisablingScript.php";
+        $script = ROOT_PATH . "models/GameCourse/Course/scripts/AutoDisablingScript.php";
         if ($endDate) {
             // Check if date is in the future
             if (date("Y-m-d H:i:s", strtotime($endDate)) > date("Y-m-d H:i:s", time()))
@@ -1444,10 +1486,10 @@ class Course
      */
     private static function validateCourse($name, $short, $color, $year, $startDate, $endDate, $isActive, $isVisible)
     {
-        self::validateName($name);
+        self::validateYear($year);
+        self::validateName($name, $year);
         self::validateShort($short);
         self::validateColor($color);
-        self::validateYear($year);
         self::validateDateTime($startDate);
         self::validateDateTime($endDate);
         self::validateStartAndEndDates($startDate, $endDate);
@@ -1461,7 +1503,7 @@ class Course
      *
      * @throws Exception
      */
-    private static function validateName($name)
+    private static function validateName($name, $year, int $courseId = null)
     {
         if (!is_string($name) || empty(trim($name)))
             throw new Exception("Course name can't be null neither empty.");
@@ -1472,6 +1514,11 @@ class Course
 
         if (iconv_strlen($name) > 100)
             throw new Exception("Course name is too long: maximum of 100 characters.");
+
+        $whereNot = [];
+        if ($courseId) $whereNot[] = ["id", $courseId];
+        if (!empty(Core::database()->select(self::TABLE_COURSE, ["name" => $name, "year" => $year], "*", null, $whereNot)))
+            throw new Exception("Duplicate course '$name' for academic year '$year'.");
     }
 
     /**
